clean up for create mode
